refactor: 쿼리키 분리
