+++ src/components/post/PostForms.tsx
@@ -1,12 +1,12 @@
 'use client';
 
-import DateInputWithIcon from './query/DateInputWithIcon';
-import TimeInputWithIcon from './query/TimeInputWithIcon';
-import { saveTags, updateImagePostId } from './query/formActions';
-import { formSchema } from './query/formSchemas';
-import { useForm } from './query/useFormMutations';
-import { usePostForm } from './query/usePostForm';
+import DateInputWithIcon from '../ui/DateInputWithIcon';
+import TimeInputWithIcon from '../ui/TimeInputWithIcon';
+import { usePostForm } from '@/hooks/byUse/usePostForm';
+import { useForm } from '@/hooks/queries/post/useFormMutations';
 import { IconPluslg } from '@/lib/icon/Icon_Plus_lg';
+import { formSchema } from '@/schemas/formSchemas';
+import { saveTags, updateImagePostId } from '@/services/server-action/formActions';
 import { useImageUploadStore } from '@/stores/post/useImageUploadStore';
 import { usePostDataStore } from '@/stores/post/usePostDataStore';
 import useBottomSheetStore from '@/stores/story/useBottomSheetStore';
