+++ src/hooks/byUse/usePostForm.ts
@@ -1,10 +1,10 @@
-import { postSchema } from '@/schemas/postSchema';
+import { formSchema } from '../../schemas/formSchemas';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
 
 export const usePostForm = () => {
   return useForm({
     mode: 'onChange',
-    resolver: zodResolver(postSchema),
+    resolver: zodResolver(formSchema),
   });
 };
